+++ examples/sqlmodel/tests/crud/test_get_one.py
@@ -38,7 +38,7 @@ async def test_get_one_relations(async_session, test_user_data, test_request, in
         "projects.company": JoinOptionModel(select=True)
     }
     exist_user_id = test_user_data[0]["id"]
-    fetched_record = await user_service.crud_get_one(test_request, exist_user_id, db_session=async_session, joins=joins)
+    fetched_record = await user_service.crud_get_one(test_request, exist_user_id, db_session=async_session, join_options=joins)
     assert fetched_record is not None
     assert len(fetched_record.projects) > 0
     assert fetched_record.projects[0].company is not None
@@ -60,7 +60,7 @@ async def test_get_one_relations_with_additional_filter_fn(async_session, test_u
         ),
     }
     exist_user_id = test_user_data[0]["id"]
-    fetched_record = await user_service.crud_get_one(test_request, exist_user_id, db_session=async_session, joins=joins)
+    fetched_record = await user_service.crud_get_one(test_request, exist_user_id, db_session=async_session, join_options=joins)
     assert fetched_record is not None
     assert fetched_record.roles is not None
     assert len(fetched_record.roles) == 1